binary(): type the dask+cupy meta so the lazy dtype matches float32 (#3508)
The _cpu_binary float32 fix and the verify_dtype=True backend tests landed
on main via the duplicate accuracy-sweep PR #3514. This change is the
remaining piece: _run_dask_cupy_binary passed meta=cupy.array(()), which
defaults to float64, so the lazy dask+cupy array advertised float64 while
_run_cupy_binary computes float32. The merged tests only check the computed
dtype (general_output_checks computes before asserting), so the mismatch
went unnoticed.

Pass meta=cupy.array((), dtype='f4') and assert the lazy dtype in
test_binary_dask_cupy. Same advertised-vs-computed class as aspect #2682 and
focal #3217. Also records the classify metadata-propagation sweep state.

+++ xrspatial/classify.py
@@ -104,7 +104,8 @@ def _run_cupy_binary(data, values):
 
 
 def _run_dask_cupy_binary(data, values_cupy):
-    out = data.map_blocks(lambda da: _run_cupy_binary(da, values_cupy), meta=cupy.array(()),
+    out = data.map_blocks(lambda da: _run_cupy_binary(da, values_cupy),
+                          meta=cupy.array((), dtype='f4'),
                           **_dask_task_name_kwargs('xrspatial.binary'))
     return out
